requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi @ file:///private/tmp/cffi-20231014-5454-1xfyd4s/cffi-1.16.0
cryptography @ file:///private/tmp/python-cryptography-20231024-6215-1s33lvq/cryptography-41.0.5
distlib==0.3.7
Django==4.2.6
django-filter==23.3
djangorestframework==3.14.0
docutils @ file:///private/tmp/docutils-20231015-5350-1xd6hrn/docutils-0.20.1
filelock==3.12.2
install==1.3.5
Markdown==3.5.1
pbr==5.11.1
pipenv==2023.10.20
platformdirs==3.11.0
pycparser @ file:///private/tmp/pycparser-20231013-5559-1dtd22y/pycparser-2.21
pyparsing @ file:///private/tmp/python-pyparsing-20231009-5290-1bhvb84/pyparsing-3.1.1
pytz @ file:///private/tmp/python-pytz-20231024-5323-3xz6j6/pytz-2023.3.post1
six==1.16.0
sqlparse==0.4.4
stevedore==5.1.0
virtualenv==20.24.6
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
